Added the UserRepository interface to the project

+++ src/main/java/Flixxer/Flixxer/Backend/DB/User.java
@@ -8,7 +8,7 @@ public class User {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int userId;
+    private long userId;
 
     @Column
     private String username;
@@ -34,11 +34,11 @@ public class User {
     public User() {
     }
 
-    public int getUserId() {
+    public long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(long userId) {
         this.userId = userId;
     }
